Call out a specific version of libusb .so file
It seems that using the current `usb-1.0` constant, rakudo/zef will fail to find libusb on a system. If you change it to the explicit file name I used, it will be found correctly. Tested against Pop! OS 22.04 LTS (which is based on Ubuntu 22.04 LTS) and Fedora 36 Workstation. My understanding is that this is a standard library naming convention in Linux, but if not it may make sense to make a Build.rakumod script that verifies that the correct *.so file can be found first.